+++ src/mdf.h
@@ -171,88 +171,5 @@ parse_mdf(std::istream &ist, bool unroll = false) {
     }
     return mdfs;
 }
-/*
-inline void
-print_mdf(ostream &ost, const pcr_copy &molecule) {
-    // For now use the depth of first pcr_copy
-    print_tsv(ost, "+" + molecule.id, molecule.depth, molecule.comment);
 
-    const vector<std::pair<int, char>> &errors = molecule.errors_so_far;
-    int size_so_far                            = 0;
-    for (const ginterval &ival : molecule.segments) {
-        string error_str = "";
-
-        for (std::pair<int, char> error : errors) {
-            if (error.first > size_so_far && error.first < size_so_far + ival.end - ival.start) {
-                error_str += (std::to_string(error.first - size_so_far) + error.second + ",");
-            }
-        }
-        if (error_str != "") {
-            error_str.pop_back();
-        }
-
-        print_tsv(ost, ival.chr, ival.start, ival.end, (ival.plus_strand ? "+" : "-"), error_str);
-        size_so_far += (ival.end - ival.start);
-    }
-}
-
-
-inline void print_mdf(ostream &ost, const string &id, const pcr_molecule &molecule, const vector< vector<std::pair<int,
-char>>> &errors_per_segment){
-
-
-    //For now use the depth of first pcr_copy
-    print_tsv(ost, "+"+id, molecule.paired[0].depth, molecule.paired[0].comment);
-
-    int interval_counter = 0;
-    for( const pcr_copy &pcp : molecule.paired){
-        int size_so_far = 0;
-        for( const ginterval &ival : pcp.segments){
-            string error_str = "";
-
-            const vector<std::pair<int, char>> &errors = errors_per_segment[interval_counter];
-
-            for(std::pair<int, char> error : errors){
-                if( error.first > size_so_far && error.first < size_so_far + ival.end - ival.start){
-                    error_str += (std::to_string(error.first-size_so_far) + error.second + ",");
-                }
-            }
-            if(error_str != ""){
-                error_str.pop_back();
-            }
-
-            print_tsv(ost, ival.chr, ival.start, ival.end, (ival.plus_strand?"+":"-"), error_str);
-            size_so_far += (ival.end-ival.start);
-            ++interval_counter;
-        }
-    }
-}
-
-inline void
-print_all_mdf(ostream &ost, const vector<pcr_copy> &molecules) {
-    for (const pcr_copy &molecule : molecules) {
-        // For now use the depth of first pcr_copy
-        print_tsv(ost, "+" + molecule.id, molecule.depth , molecule.comment);
-
-        int size_so_far = 0;
-        for (const ginterval &ival : molecule.segments) {
-            string error_str = "";
-
-            const vector<std::pair<int, char>> &errors = molecule.errors_so_far;
-
-            for (std::pair<int, char> error : errors) {
-                if (error.first > size_so_far && error.first < size_so_far + ival.end - ival.start) {
-                    error_str += (std::to_string(error.first - size_so_far) + error.second + ",");
-                }
-            }
-            if (error_str != "") {
-                error_str.pop_back();
-            }
-
-            print_tsv(ost, ival.chr, ival.start, ival.end, (ival.plus_strand ? "+" : "-"), error_str);
-            size_so_far += (ival.end - ival.start);
-        }
-    }
-}
-*/
 #endif
